+++ changelog.md
@@ -13,7 +13,10 @@
 ## Upcoming
 * Blast furnaces now try to pick up an item every 5 seconds if empty
 * Smartness now works on Fishing Rods
-* Experience and lapis are no longer consumed in creative when enchanting
+* Enchanting
+  * Lapis cost is now equal to the sum of the levels of enchantments
+  * Experience and lapis are no longer consumed in creative when enchanting
+  * Cost is rounded to the nearest integer now
 * Fixed enchanting table learning curses
 * Fixed gravity defying not reducing fall damage on servers
   * Also reduced gravity defying fall damage reduction from 4 per level to 2.5